Changing 'token' to 'password' and fixing cancelScheduledSms test
The name 'token' at setCredentials function, may confuse users with
access tokens, where there is no need to use account and password to
access.

I changed that, besides that, the cancelScheduledSms test was trying
canceling a sms scheduled to now(). That was causing failure, i added
a time to the date so the test was able to cancel it.